Move allgather Chunk into separate file
We will use this in the sparse all to all as well.
