+++ cpp/include/rapidsmpf/coll/allgather.hpp
@@ -17,8 +17,8 @@
 
 #include <rmm/cuda_stream_view.hpp>
 
+#include <rapidsmpf/coll/utils.hpp>
 #include <rapidsmpf/communicator/communicator.hpp>
-#include <rapidsmpf/error.hpp>
 #include <rapidsmpf/memory/buffer.hpp>
 #include <rapidsmpf/memory/buffer_resource.hpp>
 #include <rapidsmpf/memory/packed_data.hpp>
@@ -35,221 +35,6 @@
  */
 namespace rapidsmpf::coll {
 namespace detail {
-
-/// @brief Type alias for chunk identifiers.
-using ChunkID = std::uint64_t;
-
-/**
- * @brief Represents a data chunk in the allgather operation.
- *
- * A chunk is either a data message (in which case metadata indicates
- * how the data are to be interpreted), or a control (finish) message
- * (in which case metadata and data are empty). Chunks within a single
- * `AllGather` operation are uniquely identified by an `(id,
- * is_finish)` pair.
- */
-class Chunk {
-  private:
-    ChunkID id_;  ///< Unique chunk identifier
-    std::unique_ptr<std::vector<std::uint8_t>> metadata_;  ///< Serialized metadata
-    std::unique_ptr<Buffer> data_;  ///< Data buffer
-    std::uint64_t
-        data_size_;  ///< Size of data in bytes (maintained separately from the data
-                     ///< buffer for validation during `attach_data_buffer`)
-
-    /**
-     * @brief Construct a data chunk.
-     *
-     * @param id Unique chunk identifier.
-     * @param metadata Serialized metadata for the chunk.
-     * @param data Data buffer containing the chunk's payload.
-     *
-     * @throw std::invalid_argument If either @p metadata and @p data are null.
-     * @throw std::logic_error If one of @p metadata or @p data are null, but not both
-     * (see warning for details).
-     *
-     * @warning The caller is responsible to ensure both metadata and data are non-null,
-     * providing only one leads to an irrecoverable condition.
-     */
-    Chunk(
-        ChunkID id,
-        std::unique_ptr<std::vector<std::uint8_t>> metadata,
-        std::unique_ptr<Buffer> data
-    );
-
-    /**
-     * @brief Construct a finish marker chunk.
-     *
-     * @param id Unique chunk identifier for the finish marker.
-     *
-     * @note We use the finish marker chunk ID to encode the number of
-     * insertions on the originating rank.
-     */
-    Chunk(ChunkID id);
-
-  public:
-    /**
-     * @brief Check if the chunk is ready for processing.
-     *
-     * A chunk is ready either if it has no data buffer, or any
-     * outstanding operations on the data buffer have completed.
-     *
-     * @return True if the chunk is ready, false otherwise.
-     */
-    [[nodiscard]] bool is_ready() const noexcept;
-
-    /**
-     * @brief Return the memory type of the chunk.
-     *
-     * @return The memory type of the chunk.
-     * @note a finish chunk has memory type host.
-     */
-    [[nodiscard]] MemoryType memory_type() const noexcept;
-
-    /**
-     * @brief Check if this is a finish marker chunk.
-     *
-     * @return True if this chunk represents a finish marker, false otherwise.
-     */
-    [[nodiscard]] bool is_finish() const noexcept;
-
-    /**
-     * @brief The unique identifier of the chunk.
-     *
-     * @return The chunk's unique identifier.
-     */
-    [[nodiscard]] ChunkID id() const noexcept;
-
-    /**
-     * @brief The sequence number of the chunk.
-     *
-     * @return The sequence number portion of the chunk ID.
-     */
-    [[nodiscard]] ChunkID sequence() const noexcept;
-
-    /**
-     * @brief The origin rank of the chunk.
-     *
-     * @return The rank that originated this chunk.
-     */
-    [[nodiscard]] Rank origin() const noexcept;
-
-    /**
-     * @brief The size of the data buffer in bytes.
-     *
-     * @return The size of the chunk's data buffer.
-     */
-    [[nodiscard]] std::uint64_t data_size() const noexcept;
-
-    /**
-     * @brief The size of the metadata buffer in bytes.
-     *
-     * @return The size of the chunk's metadata.
-     */
-    [[nodiscard]] std::uint64_t metadata_size() const noexcept;
-
-    /**
-     * @brief Create a data chunk from packed data.
-     *
-     * @param sequence The sequence number for the chunk.
-     * @param origin The originating rank.
-     * @param packed_data The packed data to create the chunk from.
-     * @return A unique pointer to the created chunk.
-     */
-    [[nodiscard]] static std::unique_ptr<Chunk> from_packed_data(
-        std::uint64_t sequence, Rank origin, PackedData&& packed_data
-    );
-
-    /**
-     * @brief Create an empty finish marker chunk.
-     *
-     * @param num_local_insertions The number of data insertions on
-     * this rank.
-     * @param origin The originating rank.
-     * @return A unique pointer to the created finish marker chunk.
-     */
-    [[nodiscard]] static std::unique_ptr<Chunk> from_empty(
-        std::uint64_t num_local_insertions, Rank origin
-    );
-
-    /**
-     * @brief Release the chunk's data as PackedData.
-     *
-     * @return The chunk's data and metadata as PackedData.
-     *
-     * @throws std::logic_error if the chunk is not a data chunk.
-     *
-     * @note Behaviour is undefined if the chunk is used after being
-     * released.
-     */
-    [[nodiscard]] PackedData release();
-
-    /// @brief Number of bits used for the sequence ID in the chunk identifier.
-    static constexpr std::uint64_t ID_BITS = 38;
-    /// @brief Number of bits used for the rank in the chunk identifier.
-    static constexpr std::uint64_t RANK_BITS =
-        sizeof(ChunkID) * std::numeric_limits<unsigned char>::digits - ID_BITS;
-
-    /**
-     * @brief Create a `ChunkID` from a sequence number and origin rank.
-     *
-     * @param sequence the sequence number.
-     * @param origin the origin rank.
-     *
-     * @return The new chunk id.
-     */
-    static constexpr ChunkID chunk_id(std::uint64_t sequence, Rank origin);
-
-    /**
-     * @brief Serialize the metadata of the chunk to a byte vector.
-     *
-     * @return A vector containing the serialized chunk data.
-     */
-    [[nodiscard]] std::unique_ptr<std::vector<std::uint8_t>> serialize() const;
-
-    /**
-     * @brief Deserialize a chunk from a byte vector of its metadata.
-     *
-     * @param data The serialized chunk data.
-     * @param br Buffer resource for memory allocation.
-     * @return A unique pointer to the deserialized chunk.
-     *
-     * @note If the serialized form encodes a data chunk, this
-     * function allocates space for the data buffer.
-     */
-    [[nodiscard]] static std::unique_ptr<Chunk> deserialize(
-        std::vector<std::uint8_t>& data, BufferResource* br
-    );
-
-    /**
-     * @brief Release and return the data buffer.
-     *
-     * @return The data buffer, leaving the chunk without data.
-     */
-    [[nodiscard]] std::unique_ptr<Buffer> release_data_buffer() noexcept;
-
-    /**
-     * @brief Attach a data buffer to this chunk.
-     *
-     * @param data The data buffer to attach.
-     * @throws std::logic_error If the `data_size()` of the chunk does
-     * not match the size of the provided new data buffer, or the
-     * chunk already has a data buffer.
-     */
-    void attach_data_buffer(std::unique_ptr<Buffer> data);
-
-    /// @brief Default destructor.
-    ~Chunk() = default;
-    /// @brief Move constructor.
-    Chunk(Chunk&&) = default;
-    /// @brief Move assignment operator.
-    /// @return Moved this
-    Chunk& operator=(Chunk&&) = default;
-    /// @brief Deleted copy constructor.
-    Chunk(Chunk const&) = delete;
-    /// @brief Deleted copy assignment operator.
-    Chunk& operator=(Chunk const&) = delete;
-};
 
 /**
  * @brief A thread-safe container for managing chunks in an `AllGather`.
